apps/governance-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.2
cryptography==43.0.1
